+++ src/main/kotlin/com/jys/kotlin_practice/keycloak/KeycloakClient.kt
@@ -1,16 +1,19 @@
 package com.jys.kotlin_practice.keycloak
 
+import com.jys.kotlin_practice.account.SignupRequest
 import org.keycloak.admin.client.KeycloakBuilder
 import org.keycloak.admin.client.resource.RealmResource
+import org.springframework.http.HttpStatus
+import org.springframework.web.server.ResponseStatusException
 
 class KeycloakClient(
-    private val realm: String,
-    private val grantType: String,
-    private val authUrl: String,
-    private val clientId: String,
-    private val clientSecret: String
+    realm: String,
+    grantType: String,
+    authUrl: String,
+    clientId: String,
+    clientSecret: String
 ) {
-    fun build(): RealmResource =
+    private val realmResource: RealmResource =
         KeycloakBuilder.builder()
             .realm(realm)
             .grantType(grantType)
@@ -18,4 +21,40 @@ class KeycloakClient(
             .clientId(clientId)
             .clientSecret(clientSecret)
             .build().realm(realm)
+
+    /**
+     * keycloak 회원 생성
+     * @param request 회원 가입 요청 데이터
+     */
+    fun registerBy(request: SignupRequest) {
+        val response = realmResource.users().create(createUserBy(request))
+        if (response.status != 201) throw ResponseStatusException(HttpStatus.valueOf(response.status))
+        addRestClientRole(getByEmail(request.email).id)
+    }
+
+    /**
+     * 이메일로 keycloak 유저 조회
+     *
+     * @param email 이메일
+     */
+    fun getByEmail(email: String) =
+        realmResource.users().search(email, true).firstOrNull() ?: throw ResponseStatusException(HttpStatus.NOT_FOUND)
+
+    /**
+     * rest-client 권한 부여
+     *
+     * @param id 회원 번호
+     */
+    private fun addRestClientRole(id: String) {
+        val role = getRoleByName(KeyCloakRoles.REST_CLIENT)
+        realmResource.users().get(id).roles().realmLevel().add(listOf(role))
+    }
+
+    /**
+     * 권한 조회
+     *
+     * @param role 권한 명
+     */
+    private fun getRoleByName(role: String) = realmResource.roles()
+        .get(role).toRepresentation()
 }
